+++ tests/test_gas_boiler.py
@@ -48,8 +48,9 @@ class TestGasBoiler(TestCase):
         np.testing.assert_array_less(0.0, results["GasHeater_f713.Heat_source"])
         np.testing.assert_array_less(0.0, results["GasProducer_82ec.Gas_source_mass_flow"])
         np.testing.assert_array_less(
-            parameters["GasHeater_f713.internal_energy"]
-            * results["GasHeater_f713.GasIn.mass_flow"],
+            parameters["GasHeater_f713.energy_content"]
+            * results["GasHeater_f713.GasIn.mass_flow"]
+            / 1000.0,  # [J/kg] * [g/s] / 1000.0 = [J/s]
             results["GasHeater_f713.Heat_source"] + 1.0e-6,
         )
 
@@ -61,3 +62,9 @@ class TestGasBoiler(TestCase):
             bounds["Pipe_a7b5.GasIn.Q"][1],
             parameters["Pipe_a7b5.diameter"] ** 2 / 4 * math.pi * v_max_gas,
         )
+
+
+if __name__ == "__main__":
+
+    a = TestGasBoiler()
+    a.test_gas_boiler()